requirements.txt
streamlit==1.35.0
gspread==6.0.1
oauth2client==4.1.3
python-dotenv==1.0.1
openai==1.30.1
azure-cognitiveservices-speech==1.34.1
pandas==2.2.2
tabulate